[BZ-1086847] commit CVE-2013-4286 fix

+++ branches/JBOSSWEB_7_2_2_FINAL_BZ-1086847/src/main/java/org/apache/coyote/Request.java
@@ -288,7 +288,7 @@ public final class Request {
     }
 
 
-    public void setContentLength(int len) {
+    public void setContentLength(long len) {
         this.contentLength = len;
     }
